+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 * Captions now move down when controls are hidden ([view](https://github.com/videojs/video.js/pull/1053))
 * Added the .less source file to the distribution files ([view](https://github.com/videojs/video.js/pull/1056))
 * Changed src() to return the current selected source ([view](https://github.com/videojs/video.js/pull/968))
+* Added a grunt task for opening the next issue that needs addressing ([view](https://github.com/videojs/video.js/pull/1059))
 
 --------------------